Move the mutator controller stats reporting code into pkg/controller/mutators/
