+++ pkg/controller/mutators/stats_reporter.go
@@ -1,4 +1,4 @@
-package mutationreport
+package mutators
 
 import (
 	"context"
